requirements.txt
py-cord-dev==2.5.0
mcstatus==11.1.1
python-dotenv==1.0.1
PyYAML==6.0.1